core: refacto of STDCMGraph
